+++ src/allotropy/parsers/appbio_quantstudio/views.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from abc import ABC, abstractmethod
 from collections.abc import Iterator
-from typing import Generic, Optional, TypeVar, Union
+from typing import Generic, TypeVar
 
 from allotropy.exceptions import AllotropeConversionError
 
@@ -10,7 +10,7 @@ T = TypeVar("T")
 
 
 class ViewData(Generic[T]):
-    def __init__(self, data: dict[str, Union[ViewData[T], list[T]]]):
+    def __init__(self, data: dict[str, ViewData[T] | list[T]]):
         self.data = data
 
     def get_first_key(self) -> str:
@@ -25,7 +25,7 @@ class ViewData(Generic[T]):
             else:
                 yield [key]
 
-    def get_item(self, *keys: str) -> Union[ViewData[T], list[T]]:
+    def get_item(self, *keys: str) -> ViewData[T] | list[T]:
         if len(keys) == 0:
             return self.data[self.get_first_key()]
 
@@ -52,7 +52,7 @@ class ViewData(Generic[T]):
 
 
 class View(ABC, Generic[T]):
-    def __init__(self, sub_view: Optional[View[T]] = None):
+    def __init__(self, sub_view: View[T] | None = None):
         self.sub_view = sub_view
 
     @abstractmethod