+++ src/allotropy/parsers/beckman_vi_cell_xr/vi_cell_xr_parser.py
@@ -2,7 +2,7 @@ from __future__ import annotations
 
 from enum import Enum
 import tempfile
-from typing import Any, Optional
+from typing import Any
 import zipfile
 
 import pandas as pd
@@ -121,7 +121,7 @@ class ViCellXRParser(VendorParser):
             ),
         )
 
-    def _get_device_serial_number(self, file_info: pd.Series[Any]) -> Optional[str]:
+    def _get_device_serial_number(self, file_info: pd.Series[Any]) -> str | None:
         serial = str(file_info["serial"])
         try:
             serial_number = serial[serial.rindex(":") + 1 :].strip()