+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 ### Changed
+- Upgraded allotropy python requirement to python 10
+- Updated ASM model class typing to use or union
 
 ### Deprecated
 
@@ -28,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Global definition of calculated data documents representation
 - Update bioplex to use _get_date_time
+- Add structure for Methodical Mind
 
 ### Fixed
 - Remove duplicated ct sd and ct se calculated data documents in Quantstudio Design and Analysis adapter